requirements.txt
shodan==1.28.0
requests==2.28.1
pillow==9.4.0
rich==13.3.1
halo == 0.0.31
fire==0.5.0
python-dotenv==0.21.1
pycountry==22.3.5
clarifai_grpc==8.8.0
opencv-python==4.7
torch==2.0.1
torchvision==0.15.2
